Don't blame files/buffers outside of git
The fix for issue #30 introduced a "Not Committed Yet" annotation for
untracked files in git repositories, but unfortunately also for files
entirely outside git repositories, diff views, and other buffers where
they don't make much sense.

Flip the `check_is_ignored` test so that it only returns false if a
buffer is backed by a file that is in a git repository and is not
ignored.